Do not truncate the SQLite int to 32 bit
The integer in SQLite is a 8-byte type with variable length storage.
When an application specifies "int" or "integer" as a column type, it is
not reasonable to assume intent to truncate it to 4 bytes.  The current
behavior makes the ClickHouse SQLite integration unusable for timestamps
which are normally stored as integer.

The old behavior of truncating to narrower types when the column name
explicitly specifies something like 'smallint' is preserved.