Add ability to ATTACH tables without database layer (avoids UUID hack)
Previously every ATTACH/CREATE assumes that the table is part of
database, so it always appends the path inside this database to the
data. But, sometimes, i.e. web or s3_plain disks you amy want to specify
absolute path to the table data (where format_version.txt and all other
parts are stored). This can be useful if you want to share the data via
s3 bucket or via http server.

Previously it was possible with ATTACH x UUID 'exact-uuid', but that was
very annoying, i.e. you need to know it, plus, you cannot create two
such tables on one server (of course you can workaround this with
clickhouse-local).

TL;DR;

I considiered multiple ideas on implementing this, but after all, this
looks like the best approach (even though it has extra condition in the
MergeTree code).

- table_endpoint

  Use table_endpoint over endpoint for configuring disks - this will
  require extra code in disks and handling it for MergeTree anyway, and
  it is not better then the MergeTree setting.

  It could be done as table_endpoint flag, though it will still require
  handling and passing it through all the layers to the MergeTree.

  Plus, I this is the table of the property, not the disk.

- special database

  This is also a good option, but it will allow you to attach such
  tables into existing database, which can be useful as well (I guess).

- separate disk

  It could be implemented as a disk wrapper, but, it will require to accept the
  relative path for the table and strip it out, does not sounds better to me