Remove unnecessary include
If no destructor is added, an implicit destructor is automatically
generated in the header file for every translation unit including
the header. Since it's the destructor the one that needs to know
the size of what std::unique_ptr points to in order to know how
to destroy it, the best thing we can do is to define a default
destructor in the .cpp to prevent any automatic destructors
being generated via header include.