Consider UDF non-deterministic.
Extend the HasNonDeterministicFunctionsMatcher to check also UDFs instead of only user defined functions.

* QueryResultCache.cpp: (HasNonDeterministicFunctionsMatcher::visit) : Try harder with UserDefinedSQLFunctionFactory and
UserDefinedExecutableFunctionFactory if the functon is not deterministic.

Fixes #77553

Todo: The two new tryGet functions need to take a lock so this change could potentially impact the command processing
when the server is used concurrently and the users are sending SQLs with many UDFs.