tests: fix 03293_tty_binary_confirmation flakiness
The problem is that it is likely interfer with progress and never
receive the match that expect(1) expectds

I.e. here [1] it does not match first "PAR", while there is binary in
the output, so it is definitelly should be there, and it is release
build, timeout is not possible there.

  [1]: https://s3.amazonaws.com/clickhouse-test-reports/json.html?REF=master&sha=b928c8e96e24f0bd0a69743d03b972c900813b30&name_0=MasterCI&name_1=Stateless%20tests%20%28release%29