Fix skip indices are not used with OR predicates
The underlying issue is that every index is checked separately [here](https://github.com/ClickHouse/ClickHouse/blob/d2b015ee5f5d9890c6c1e3a231bc07d69d664b97/src/Processors/QueryPlan/ReadFromMergeTree.cpp#L1746) if it might be useful in the query. The issue arises when there is an `OR` predicate in the filter query because a simple state about RPN evaluates to true or unknown is not enough. This is a corner case only for the `OR` predicate.
If we replace the `OR` predicate the query from #75228 with
`NOT(colA != 'aa' AND colB != 'bb')` (which is logically the same as `OR`)
as `EXPLAIN indexes=1 SELECT * FROM or_test WHERE id IN (0,1,2,3,4) and NOT(colA!='aa' AND colB!='bb')`,
the skip indices would be used as expected:
```
EXPLAIN indexes = 1
SELECT *
FROM or_test
WHERE (id IN (0, 1, 2, 3, 4)) AND (NOT ((colA != 'aa') AND (colB != 'bb')))

Query id: 0a5ff444-0274-47d1-b790-fb763dc6ad73

    ┌─explain─────────────────────────────────────────┐
 1. │ Expression ((Project names + Projection))       │
 2. │   Expression                                    │
 3. │     ReadFromMergeTree (gh_75228.or_test)        │
 4. │     Indexes:                                    │
 5. │       PrimaryKey                                │
 6. │         Keys:                                   │
 7. │           id                                    │
 8. │         Condition: (id in 5-element set)        │
 9. │         Parts: 1/1                              │
10. │         Granules: 5/5                           │
11. │       Skip                                      │
12. │         Name: colA_idx                          │
13. │         Description: bloom_filter GRANULARITY 1 │
14. │         Parts: 1/1                              │
15. │         Granules: 5/5                           │
16. │       Skip                                      │
17. │         Name: colB_idx                          │
18. │         Description: bloom_filter GRANULARITY 1 │
19. │         Parts: 1/1                              │
20. │         Granules: 5/5                           │
    └─────────────────────────────────────────────────┘
```

To fix it, a new method `isIndexUseful` has been added to
`MergeTreeIndexConditionBloomFilter` which decides the usefulness of the
index based on multiple states.

Fix #75228