ci: fix cache for clang-tidy
First of the path is wrong, it should be "ccache/clang-tidy-cache"

Secondly, why do we set CTCACHE_S3_NO_CREDENTIALS this looks wrong, it
will not update the cache. Note, that this is done only under "not
local", so should be OK to enable it.

v2: rebased patches to ensure that cache for clang-tidy works!