Fix dumping network profile events for HTTP protocol with compression enabled
The problem is that in this case the network profile events
(NetworkSendElapsedMicroseconds/NetworkSendBytes) will be incremented
when the entry to query_log had been flushed already.

Note, the drift should not be more then the buffer size (usually around
1MiB).

Als note, that we do not flush the buffers in case of exception, since
this is too tricky.

v2: do not call callback on exception, since it is too tricky