ci: add symbolizer to clickhouse/integration-tests-runner
Yes, server does not need it since it uses clickhouse/integration-test,
but the client uses clickhouse/integration-tests-runner, and the client
also may have races.

So I've switched the image to be based on clickhouse/test-base, this
image is ~3GiB in size, but the runner image itself is also around
~3GiB, so let's see what the final size will be.

But anyway I think it is not that big deal.