ci: disable async_insert for sending system.*_log to ci-logs cluster
I found this on CI [1]:

    2025.03.25 21:30:01.453922 [ 129437 ] {} <Error> system.error_log_sender.DistributedInsertQueue.default: Code: 159. DB::Exception: Received from kng4alm55c.us-east-2.aws.clickhouse-staging.com:9440. DB::Exception: Wait for async insert timeout (120000 ms) exceeded). Stack trace:

  [1]: https://github.com/ClickHouse/ClickHouse/pull/78266#issuecomment-2770842740

And async_insert is useless there, since the batches should be big enough.

Note, that I thought about implementing some hints for background
Distributed INSERTs, relying on async_insert_max_data_size, but the
problem is that it will rely on client setting value, not the server.