fix clickhouse server startup issue (macOS debug build)
The clickhouse server cannot start due to allocation of `ThreadStack`.
In case of a debug build, we were increasing the size by one and this
led to having data size not aligned with the data alignment size.

Instead of increase the data size by one, we can increase by the page
size.