client: add ability to format query for editor (via M-F)
The formatting will multi-line if the query is multi-line itself, and
online otherwise.

You can use EDITOR=cat to just format queries in the client instead of
editing them.

Note, that the old shortcut (M-E) has not been replaced, and will not
format query, since sometimes in case of queries that does not fit into
a screen you don't want this

v2: use C-; over C-E (it is readline jump to end of line)
v3: use M-F (F stands for format) over M-E (C-; is tricky since it does not have normal representation as C-[a-z] has)