Fix excessive number of files (leads to excessive memory usage) for external ORDER BY
In #72598 the behavior of `max_bytes_before_external_sort` has been
changed, it has been switched to rely on the total memory usage of the
query instead of the size of a sorting data.

But, the problem with this approach is that you can have a lot of memory
used outside of sorting (FINAL, aggregation), and spilling sorting
blocks to disk will not help reduce memory usage obviously, but it will
make things even worse, because it will create lots of small files, that
will require a lot of memory to merge them.

So this patch reverts the behavior of `max_bytes_before_external_sort`
back, now it will rely on the sorting block size.