Fix memory accounting for Iceberg metadata files cache
The problem is that the JSON object takes more bytes in memory, then a
plain string (via side channels the overhead is around 10-20x, so if you
have 500MiB of cache, in reality it can take ~10GiB of RAM)

So let's simply storing the std::string itself, yes this will be
slightly less optimal, but I doubt that the difference will be
significant.

As a long-term we can fix this by using some intrusive parser, or JSON
parser that allows to pass the allocator (AFAICS both variants can be
implemented with rapidjson), but for now, let's keep it simple to
backport this change.