chdig: fix dependencies from rt/util (from libc crate) and also gcc_s/unwind
The problem is that libc explicitly links with all libc related
libraries, but in ClickHouse we knows better what to link with.
So let's create dummy libraries to fix the dependencies.

Note, that in theory this could be fixed by using musl target, but
changing rust target for specific cmake target is not implemented in
corrosion [1].

  [1]: https://github.com/corrosion-rs/corrosion/issues/570