Count consumed messages manually to avoid depending on previous committed offset
We had an off-by-one error when calculating the intent size. I also discovered that when we read the offset from kafka on table creation, the intent size was also wrongly calculated. As a result I changed to logic to manually count the consumed message to ensure intent size is correct.