Substitute pointer with shared pointer for index in context
This solves the issue that the index are stored in ReadFromMergeTree class that is destroyed with the planner, before the pipeline functions
execute.
However, this is not a final solution because it duplicates the index class which could become a problem.
The final solution involves using the same shared pointer to avoid duplication, but that implies multiple modifications in the
ReadFromMergeTree class.
Remember that the shared_ptr is not thread safe, which is a problem and needs explicit handling in the destruction to ensure the proper
reference counter.
TODO: Once everything works, remember to implement this properly without supplications