Avoid write with append mode requirement for MergeTree transactions
Note that for i.e. remote storages it is not optimal, since it will add
one more blob, while for local disks it does make sense to write with
append and we can have a separate path for local disks, but,
transactions are experimental, let's not introduce code that will not be
tested.

Also I've tried couple of other ideas, but the interface became more
complex, and I am not sure that it worth it for experimental
functionality.