Cache a shared vector of parts in MergeTreeData.
Even using a shared lock when acquiring the vector of parts is limited
in value, since I still observe quite a bit of time spent copying the
parts vector in the `getDataPartsVectorForInternalUsage` function. Given
this list of parts is already getting filtered later in the query
planning phase, it seems that sharing a copy of the vector across all
concurrently running SELECT queries makes sense.

The implementation of this caching depends on the previous commit, which
brings the option to either "read lock" or "write lock" the merge tree
data. In the case of a write lock, we know that this implies the set of
parts may be getting mutated, and as such it would invalidate our cached
list of parts. So, in the destructor for this lock, we recompute the
cached list.

This causes the "read vs write" lock to have a more important
difference, as a "write lock" is no longer const, and as such cannot be
called from const functions. This is why quite a few callsites needed
modified in this commit.

This brings a significant improvement to query planning performance,
particularly for tables with many parts, or where many queries are
concurrently accessing a table.