Generalize implementation considering index granularity.
This implementation considers index granularity which by default is >
1.
In that case the attempt to read the index too many times will not
scale. SO, our solution implies calling the matches only once with the
ranges as arguments.
This approach is specially good when there are few ranges to match in
a single index with big granularity.

TODO: After a profile I expect to see more cost reading the index with
granularity = 64 (default) But we already have a solution for that ;)

In this commit I also generalize the implementation to use formal data
structures and remove some micro optimizations that didn't give any
significant performance improvement.