Allow taking a shared lock of parts in MergeTreeData
In one of our internal clusters (at Cloudflare), we encountered a
situation where the majority of leaf query duration was spent waiting to
acquire the parts lock. In particular when calling
`MergeTreeData::getStorageSnapshot`.

This particular cluster has a table design which leads to quite a high
part count, which is problematic, as some paths which hold the parts
lock have to iterate over all paths. Eventually acquiring the lock
becomes more of a bottleneck than the actual work being done once the
lock is acquired.

My solution was to make it so that read-only operations over the parts
list can use a shared lock, allowing multiple entrants to hold the lock
simultaneously. This completely resolved the contention problem for us,
and *significantly* improved query performance for this workload,
without showing any negative impact to more conventional datasets.

I've chosen to only make use of this shared locking in a few code paths
where I already observed contention being a problem, but I do believe a
shared lock could be acquired in many functions here. My preference is
to keep the diff small.