Use inplace Dag function replacement.
This simplifies a lot of code by replacing the node content inplace.
Ad the node address is the same we can avoid the full redirection step and replacement we used to do.
However there is a small hack because we add a new node to pop it immediately, but so far that is not an issue.