Fix possible inconsistency for dictionaries with update_field on MEMORY_LIMIT_EXCEEDED errors
Before this patch it was possible that the dictionary with
`update_field` can be left in a inconsistent state after
`MEMORY_LIMIT_EXCEEDED` error during update (some columns was updated -
new filter has been applied and new values added, while others does not).

After this patch this block will not be cached on update, so reuse is
not possible, and we cannot simply cut the column to an old value since
we applies filters at first, but, since IColumn::filter() return new
column, so this should not increase memory usage (except for the time
when `mergeBlockWithPipe` is running).