ci: sanity checks for cherry pick script against empty responses from GitHub
Today there was one cherry-pick run that closes few PRs with a comment
"The release branch 25.7 for the cherry-pick doesn't have an opened PR,
closing this PR." even though the release PR was never closed.

But the script assumed so [1]:

    2025-08-20T14:11:09.0957244Z INFO:root:An opened release PR `25.7` for cherry-pick PR https://github.com/ClickHouse/ClickHouse/pull/85490 is not found, going to close it
    ...
    2025-08-20T14:11:48.7892801Z INFO:root:Active releases: 24.8, 25.3, 25.5, 25.6, 25.7
    ...
    2025-08-20T14:13:41.8261936Z INFO:root:The cherry-pick PR #85490 for PR #85038 is discarded

  [1]: https://github.com/ClickHouse/ClickHouse/actions/runs/17100818473/job/48496622985

Which reminds me that recently GitHub returns zero results for issue/PRs
searches, so maybe this happened again? Including API?

This is just a crazy idea.