Fix corner case with index boundaries.
This fixes the introduction of out of boundaries values for filtering in the index function.
The code is totally defensive now by using a proper number of assertions and introducing the tight ranges for filtering.
Any out of bound issue will be handled properly in release code (and produce correct results), but will trigger some assertion in debug
mode.