Fix leaking of MergesMutationsMemoryTracking due to Buffer tables
This place forgot to call
background_memory_tracker.adjustOnBackgroundTaskEnd(), like we do for
merges/mutations and this may lead to significant drift in
MergesMutationsMemoryTracking, which later will lead to TOO_MANY_PARTS,
because merges will stop.

At first I actually added it (adjustOnBackgroundTaskEnd()) but later
decided that it is wrong, merges/mutations should not take care about
Buffer flushes.

Note, maybe it will be a good idea to have a RAII for return value of
createForBackgroundProcess() that will take care of
adjustOnBackgroundTaskEnd(), but but this is not can be tricky since you
need to make sure that there are no users of it (i.e.
ThreadGroupSwitcher already finished).