Fix crc32c build on older ARM CPUs
crc32c_arm64.cc uses the vmull_p64 instruction, which is only available when
building with the `aes` flag, which is included among others under the `crypto`
flag. We only add the `crypto` flag when building on modern ARM CPUs, so we
can't build that file on older ARM CPUs.