Add support for user names in config containing a dot
For correct config parsing, the dots in the user names have been escaped
for some time [1]. This caused a confusion for some users. This commit
adds optional de-escaping to the config parsing process. In order to
keep the behaviour backwards-compatible, it's been hidden behind the new
config entry `<escape_dot_in_user_name>` which is turned on by default.

Fixes #82026

[1] https://github.com/ClickHouse/ClickHouse/pull/58354