Support move/tag action for processed S3/Azure queue objects
If action is set to move, objects can be moved to a different path prefix or another bucket/container.
Move requires additional table settings such as new path prefix, or bucket URL and credentials in case of S3, or connection string and container name in case of Azure.

Tag requires addtional settings to pass tag key and value.