feat: Support exception tagging for HTTP response
Fixes: https://github.com/ClickHouse/ClickHouse/issues/75175

This PR adds consistent and reliable way to return exceptions occured in
mid-stream during serving HTTP request.

It implements the proposal agreed on the attached issue. Everytime the
exception occurs, it will be written in following form (which makes it
easier for clients to parse the exception message reliably)

```
HTTP/1.1 200 OK↵
X-ClickHouse-Exception-Tag: PU1FNUFH98↵
↵
bodybodybodybody↵
__exception__PU1FNUFH98↵
Big bam accrued right while reading the data↵
45 PU1FNUFH98__exception__↵
```

Here 45 is the size of the Big bam accrued right while reading the data↵ message, and PU1FNUFH98 is the exception token.

Also this behavior change comes into play only when the settings
`http_exception_tagging` is 1 (default 0)
Otherwise it's a no-op

Signed-off-by: Kaviraj <kavirajkanagaraj@gmail.com>