*: avoid use of unqualified pow().
libc implementations *may* include pow() in the global namespace when we
include cmath, but this isn't guaranteed. To ensure portability with libc
implementations that don't place pow() in the global namespace, replace
unqualified pow() with std::pow(). Also explicitly include cmath where used so
that we don't rely on transitive includes.