Text index: add array support for `hasAny/AllTokens`
Previously, only a string column was supported as an input to the
`hasAnyTokens` and `hasAllTokens` functions. However, both `Array` and
`Map` (via `mapKeys` and `mapValues`) input columns are already
supported by the text index.
Since direct read from the text index is supported only by a handful
functions, it was not possible to support the direct read for `Array`
and `Map` columns.

This change adds support for an array of string column as an input to both
`hasAnyTokens` and `hasAnyAllTokens` functions which led to having a
direct read support for `Array` and `Map`.

NOTE: `Map` itself is not directly supported as an input, instead
`mapKeys` or `mapValues` functions should be a wrapper to `Map`.