Introduce per-table cache of ColumnsDescription for parts to reduce memory usage
I've seen a setup where 7GiB (and this is jemalloc statistics w/o memory
fragmentation) out of 90 was used by the ColumnsDescription, and it is
only for 20K parts (but there are 6.5e6 parts columns).

Also I've seen that with stopped merges, due to metric_log, the server
may consume ~512MiB for a couple of hours with doing almost nothing.