Make parsing of XRay instrumentation single-threaded
Turns out that it's dangerous to use a different LLVMSymbolizer
per thread because it consumes a lot of memory. When optimizing
so that each thread takes a function to symbolize, I realized that
when all cores were symbolizing at the same time the AWS dev machine
ended up unresponsive. There are some workarounds:

1. call LLVMSymbolizer.flush() once in a while to drop the cache.
   Take into account that calling pruneCache() is not enough because
   there are some other internal caches that use a lot of memory.
2. Use a single LLVMSymbolizer and a mutex. This makes the code more
   complex since there are still several threads involved. The worst
   part is that there is no performance gain since the background
   threads have lower priority than the one where the query is executed.

In summary, it's difficult to predict how much memory LLVMSymbolizer
ends up using, but the theory is that each each will use at least the
DWARF information chunk from the binary, which for a debug buidl is big.

In conclusion, I don't think having parallel threads working on it is
worth the extra complexity and the danger to leave the machine without
memory to work with.