Fix sparseGram tokenizer bug in hasAny/AllTokens and tokens functions
The bug is only limited to the `sparseGram` tokenizer. The sparseGram
tokenizer stores an internal state which is modified during the execution.

This led an error when the functions are running concurrently, because
the internal state of the sparseGram tokenizer is not thread-safe.

This change fixes that by cloning the sparseGram tokenizer for each
function call to avoid sharing an internal state between runs.