Use substrings to find instrumentation points
Fix setting instrumentation points on templated functions.

This simplifies quite a bit the code and makes it less error prone
to C++ parsing shenanigans.
Since we're now showing the complete symbol name both in
system.instrumentation and in system.symbols, we expect the users
to use a name as close as possible to the complete one.