Add the ability retry certain errors in distributed queries
As ClickHouse clusters scale and you begin to have multiple
shards with many replicas, the probability of 1 of those nodes
being saturated, busy or not in a condition to serve queries
increases.

This can often mean that a Distributed query say across dozens
of shards, might fail just because of a single node. This means
all the work and resources we've used on other shards is effectively
thrown away, and return an error to the client, which then can choose
to retry the whole query if that want to.

This is quite wasteful, and can often lead to cascading values if
clusters are busy.

This commit adds a new setting:

`distributed_shard_retry_count` which defaults to 0 to effectively turn
retries off for backwards compatible behaviour.

If this setting is > 0 and a leaf node in a distributed query returns
an error, we will retry the query on a different leaf node.

Co-authored-by: Christian Cendres
Co-authored-by: Anton Kozlov