Bug fix: handle implicit conversion from string in logical expression optimizer
The pattern of `col = X and col = Y` is assumed to yield to
`ALWAYS_FALSE` which is the true for almost all cases.
However, in case of an implicit conversion from a string to an integer
or a boolean that assumption drops since `col = X and col = 'X'` should be same.

Fixes #89803