tests/integration: use more lightweight backtrace capturing
There are reports when test_s3_plain_rewritable hungs, interesting logs
there are:

    [2025-11-21 14:32:04] 2025-11-21 13:51:09.171000 [ 503 ] WARNING : Force kill clickhouse in stop_clickhouse. ps:10 (cluster.py:4868, stop_clickhouse)
    [2025-11-21 14:32:04] 2025-11-21 13:56:09.526000 [ 503 ] WARNING : Stop ClickHouse raised an error Command '['docker', 'exec', '-u', 'root', 'roottests3plainrewritable-gw3-node1-1', 'bash', '-c', "gdb -batch -ex 'thread apply all bt full' -p 9 > /var/log/clickhouse-server/stdout.log"]' timed out after 300 seconds (cluster.py:4889, stop_clickhouse)

But, at the same time on the server:

   2025.11.21 13:50:10.910910 [ 15 ] {} <Trace> BaseDaemon: Received signal -2
   2025.11.21 13:50:10.910942 [ 15 ] {} <Information> BaseDaemon: Stop SignalListener thread

And gdb log:

    Thread 1 (process 9 "clickhouse"):
    [Inferior 1 (process 9) detached]

So apparently it hunged somewhere in destructors.

Though this is very unlikely that gdb cannot collect stacktrace in 5
minutes, although when I tried I found that full bt can take 4+ minutes,
but AFAIR it was printing at least something in the file (i.e. first few
frames).