Disable enable_shared_storage_snapshot_in_query (leads to memory leaks)
Some functions stores a copy of `ContextPtr` (not a `weak_ptr` as it
should be, but `shared_ptr`), i.e. [`modulo`](https://github.com/ClickHouse/ClickHouse/blob/407e33a47cf707155d1c4f3f12d2bb2291c80009/src/Functions/FunctionBinaryArithmetic.h#L827),
but this `IFunction` itself can be stored in `KeyDescription`, which is
part of `StorageInMemoryMetadata`.

And with metadata cache (`enable_shared_storage_snapshot_in_query=1`)
this will create memory leaks (since this `Context` will never be
destroyed)

As @Michicosun has been told, this cache should never be part of
`Context` (but we abuse it often).

(Another issue is `IFunction` holds a `ContextPtr` not a
`ContextWeakPtr`, but this is another story)

Anyway, let's disable it for now (backported to 25.11 is a must).