Reduce S3 listing based on _path values from filter
In order to push down _path values specified in query filter,
these values should be extracted from the predicate and if found,
filtered by the glob expression to avoid scanning unallowed data.

For the extraction part, `evaluateExpressionOverConstantCondition`
(which is already used for extracting sharding key values from the
predicate to skip unused shards) seems to be a good fit, so its
logic is extended to handle all FutureSet types.

Co-authored-by: Kseniia Sumarokova <sumarokovakseniia@gmail.com>