Fix `WITH TIES` logic
Fixes the `previous_row_chunk` assignment logic to ensure
that it doesn't have a value if we still haven't finished
our normal `limit + offset` processing.

That avoids making `splitChunk` break our logic as it expects
a value for `previous_row_chunk` only if we are finished with
output-ing `limit + offset`.