Remove ContextPtr from Functions (part1)
Context holds a lot, that may include some caches, by holding a copy to
ContextPtr the lifetime of this caches will be extended.

Most of the funtions do not need copy of ContextPtr, they can either
copy required prats from it or use WithContext (which holds only weak
pointer).

Though note, that the function may be holded in `KeyDescription`
(MergeTree), but it should not be called, so it should be OK to replace
`ContextPtr` with `WithContext`.

This patch does two things:
- adds a style check to avoid such problems in future (with list of
  exceptions, some of which is easy to fix, others are OK as is)
- obtain required data from ContextPtr instead of holding copy of it in
  FunctionsConversion (and in a few more)