Fix possible crash in aggregate functions after MEMORY_LIMIT_EXCEEDED
The problem is when ColumnAggregateFunction is reused after exception,
this is the case of async_insert=1, where we are trying to load each
INSERT batch, and if it fails, continue with other batches.

The problem was that once ColumnAggregateFunction::ensureOwnership()
throws, it leaves the column in a broken state, since all aggregation
states up to rollback_pos will be broken.

The fix is simple - just copy them from source column again.

And add a unit test.