Make query memory usage detection for spilling to disk during aggregation/sorting more robust
It is possible for getCurrentQueryMemoryUsage() to return memory usage
not for query, but for i.e. server, and this will make calculations
completely wrong.

The is no issue in upstream, since in 25.11 in accidentally fixed [1]

  [1]: https://github.com/ClickHouse/ClickHouse/pull/83277/files/38a41d92d7a64839dca907e4389374755600d88b#diff-b5b5b5decb56b19cb06f248a7646d9ef542019b4ec3b01d8ce7dccf93fd54750R190

But, it is only in 25.11, we need to backport this to earlier releases
with ratio for spilling to disk.