Write symbolized jemalloc profiles
It was inconvenience for a long time that you need the binary for
jemalloc heap profiler (different architecture, different versions and
so on)

This patch will write second file with a `.symbolized` prefix, for which
you do not need binary anymore for `jeprof`, and you can simply run,
i.e. `jeprof --text heap.0.symbolized`

Internally `jeprof` supports symbolized profiles, you just need to
provide a map from addresses to symbols (with some quirks) and `jeprof`
will read it out of the box.

Note, I've though about adding special hooks for jemalloc, but it is
tricky, since you cannot everything inside generating this profile, so
this implementation simply reads the file written my jemalloc, and
creates a new profile with a `.symbolized` prefix.