Fix leaving possibly broken distributed connections in the pool
Previous fix (#92807) introduce another issue, due to which the connection for
distributed queries can be preserved in the pool even after some
exception (usually it is socket timeout), which leads to a reconnect in
the next query, or even query failure.

The problem was that `RemoteQueryExecutor` uses `finished` flag to
decide should it call `finish()` again in `RemoteSource`, but this
`finished` flag is also used in dtor of `RemoteQueryExecutor`, and if
the query is not finished it is aborted, which does not work with this
new logic.

Let's use `got_exception_from_replica` instead, which does not affect
dtor. And it is OK to call `finish()` again in `RemoteSource`, since
subsequent will be no-op.